Adjust ATCamera component state mappings
